### 2024-04-11 #6606 @editorialbot
[PRE REVIEW]: orbitize! v3: Orbit-fitting for the High-contrast Imaging Community
**Submitting author:** @sblunt (<a href="http://orcid.org/0000-0002-3199-2888">Sarah Blunt</a>)
**Repository:** https://github.com/sblunt/orbitize
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v3.0.0
