### 2024-05-16 #6756 @sblunt
Thanks for reviewing everyone!! I just wanted to make you aware of something— I’m currently locked out of my pypi account, so the version of orbitize on pypi is not 3.0.0. I’ll update the pypi release as soon as I can regain access to my account!
