### 2024-05-30 0b8f27b
add note about installing from git
and a little snark
