### 2024-06-03 #6756 @DTCupcakes
I've noticed that there's some authors in the author list of the [original AJ paper](https://ui.adsabs.harvard.edu/abs/2020AJ....159...89B/abstract) that are missing from this paper. Specifically
- Isabel Angelo
- Henry Ngo
- James Graham
- Logan Pearce
- Malena Rice
Is there a reason why this is the case?
