### 2024-06-03 #6756 @DTCupcakes
When running the first cell in the [HGCA tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/HGCA_tutorial.html) I get an "InconsistentTableError". Is this an issue with orbitize? I'm not familiar with the Hipparcos-Gaia backend.
