### 2024-06-11 #6756 @sblunt
Hey @DTCupcakes! Thanks for reviewing for us. Two things:
1. This JOSS paper is meant to cover additions to the code between v1 and v3, while the original AJ paper covered additions to the code up to v1. So, the folks who you listed contributed to v1 but did not contribute subsequently.
2. Do you have v3 installed? I'm having this super annoying issue where I'm locked out of my pypi account and can't make v3 pip installable, but if you follow the [most up-to-date installation instructions](https://coolstars22.github.io/program.html), you should be able to install v3. If you do have v3 installed and you're still getting the error, let me know!
