### 2024-06-17 #6756 @sblunt
Hm, ok! I'm not sure. I can't reproduce the error on my machine. Two follow up questions:

1) Did you follow the instructions to obtain the data ([this section of the tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/HGCA_tutorial.html#Obtain-the-necessary-data)), and did you run into any issues with that?

2) Can you send the full error?
