### 2024-07-11 #6756 @DTCupcakes
@sblunt This is the only tutorial I'm having issues with. It's not clear to me if the data is already included in the example data folder included with orbitize or if I have to add it externally.
