### 2024-07-11 #6756 @DTCupcakes
Here's the full error:
`InconsistentTableError                    Traceback (most recent call last)
[/usr/lib/python3/dist-packages/astropy/io/ascii/ui.py](http://localhost:8888/usr/lib/python3/dist-packages/astropy/io/ascii/ui.py) in _guess(table, read_kwargs, format, fast_reader)
    521             reader = get_reader(**read_kwargs)
--> 522             dat = reader.read(table)
    523             _read_trace.append({'kwargs': copy.deepcopy(read_kwargs),

[/usr/lib/python3/dist-packages/astropy/io/ascii/core.py](http://localhost:8888/usr/lib/python3/dist-packages/astropy/io/ascii/core.py) in read(self, table)
   1368
-> 1369                     raise InconsistentTableError(errmsg)
   1370

InconsistentTableError: Number of header columns (13) inconsistent with data columns (2) at data line 0
Header values: ['Target,', 'ra[rad],', 'dec[rad],', 'ra[h:m:s],', 'dec[d:m:s],', 'ObservationTimeAtGaia[UTC],', 'CcdRow[1-7],', 'zetaFieldAngle[rad],', 'scanAngle[rad],', 'Fov[FovP=preceding[/FovF](http://localhost:8888/FovF)=following],', 'parallaxFactorAlongScan,', 'parallaxFactorAcrossScan,', 'ObservationTimeAtBarycentre[BarycentricJulianDateInTCB]']
Data values: ['bet', 'Pic,1.5153157780177544,-0.8912787619608181,05:47:17.088,-51:03:59.441,2014-09-24T03:17:43.690,1,0.003366090442748648,-2.300433958482974,FoVP,-0.7177768121570808,0.716584607472628,2456924.6385198794']

InconsistentTableError                    Traceback (most recent call last)
[/tmp/ipykernel_1052549/2860455245.py](http://localhost:8888/tmp/ipykernel_1052549/2860455245.py) in <module>
      9 # we're just going to assume one planet for simplicity
     10 hipparcos_lnprob = hipparcos.HipparcosLogProb(iad_filepath, 27321, 1)
---> 11 hgca_lnprob = gaia.HGCALogProb(27321, hipparcos_lnprob, gost_filepath)

[~/dev/orbitize/orbitize/gaia.py](http://localhost:8888/home/trevascus/dev/orbitize/orbitize/gaia.py) in __init__(self, hip_id, hiplogprob, gost_filepath, hgca_filepath)
    296         self.gaia_epoch_dec = entry["epoch_dec_gaia"][0]
    297         # read in the GOST file to get the estimated Gaia epochs and scan angles
--> 298         gost_dat = read(gost_filepath, converters={"*": [int, float, bytes]})
    299         self.gaia_epoch = time.Time(
    300             gost_dat["ObservationTimeAtGaia[UTC]"]

[/usr/lib/python3/dist-packages/astropy/io/ascii/ui.py](http://localhost:8888/usr/lib/python3/dist-packages/astropy/io/ascii/ui.py) in read(table, guess, **kwargs)
    339         # through below to the non-guess way so that any problems result in a
    340         # more useful traceback.
--> 341         dat = _guess(table, new_kwargs, format, fast_reader)
    342         if dat is None:
    343             guess = False

[/usr/lib/python3/dist-packages/astropy/io/ascii/ui.py](http://localhost:8888/usr/lib/python3/dist-packages/astropy/io/ascii/ui.py) in _guess(table, read_kwargs, format, fast_reader)
    551                    '************************************************************************']
    552             lines.extend(msg)
--> 553             raise core.InconsistentTableError('\n'.join(lines))
    554
    555

InconsistentTableError:
ERROR: Unable to guess table format with the guesses listed below:
Reader:Ecsv converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:FixedWidthTwoLine converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:RST converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:FastRdb converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Rdb converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:FastTab converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Tab converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Cds converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Mrt converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Daophot converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:SExtractor converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Ipac converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:Latex converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:AASTex converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')] strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastCommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:CommentedHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastBasic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:FastNoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '|' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ',' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: ' ' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: '"' strict_names: True
Reader:NoHeader converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} delimiter: '\\s' fast_reader: {'enable': True} fill_values: [('', '0')] quotechar: "'" strict_names: True
Reader:Basic converters: {'*': [<class 'int'>, <class 'float'>, <class 'bytes'>]} fast_reader: {'enable': True} fill_values: [('', '0')]

************************************************************************
** ERROR: Unable to guess table format with the guesses listed above. **
**                                                                    **
** To figure out why the table did not read, use guess=False and      **
** fast_reader=False, along with any appropriate arguments to read(). **
** In particular specify the format and any known attributes like the **
** delimiter.                                                         **
************************************************************************`
