### 2024-07-19 71569d6
Add method to system object to plot radec astrometry with error bars
