### 2024-07-19 33b499d
Add plotting of seppa data (in radec) with errors
