### 2024-07-19 efdd2e4
Add axis labels to astrometry plot and set axis limits so star is centred
