### 2024-07-19 3778043
Add astrometry plotting to init method of system class
