+++ orbitize/kepler.py
@@ -1,6 +1,7 @@
 """
 This module solves for the orbit of the planet given Keplerian parameters.
 """
+
 import numpy as np
 import astropy.units as u
 import astropy.constants as consts
@@ -13,30 +14,29 @@ if cext:
 if cuda_ext:
     # Configure GPU context for CUDA accelerated compute
     from orbitize import gpu_context
+
     kep_gpu_ctx = gpu_context.gpu_context()
+
 
 def tau_to_manom(date, sma, mtot, tau, tau_ref_epoch):
     """
     Gets the mean anomlay
-    
+
     Args:
         date (float or np.array): MJD
         sma (float): semi major axis (AU)
         mtot (float): total mass (M_sun)
         tau (float): epoch of periastron, in units of the orbital period
         tau_ref_epoch (float): reference epoch for tau
-        
+
     Returns:
         float or np.array: mean anomaly on that date [0, 2pi)
     """
 
-    period = np.sqrt(
-        4 * np.pi**2.0 * (sma * u.AU)**3 /
-        (consts.G * (mtot * u.Msun))
-    )
+    period = np.sqrt(4 * np.pi**2.0 * (sma * u.AU) ** 3 / (consts.G * (mtot * u.Msun)))
     period = period.to(u.day).value
 
-    frac_date = (date - tau_ref_epoch)/period
+    frac_date = (date - tau_ref_epoch) / period
     frac_date %= 1
 
     mean_anom = (frac_date - tau) * 2 * np.pi
@@ -44,11 +44,24 @@ def tau_to_manom(date, sma, mtot, tau, tau_ref_epoch):
 
     return mean_anom
 
-def times2trueanom_and_eccanom(sma, epochs, mtot, ecc, tau, tau_ref_epoch=58849, tolerance=1e-9, max_iter=100, use_c=True, use_gpu=False ):
+
+def times2trueanom_and_eccanom(
+    sma,
+    epochs,
+    mtot,
+    ecc,
+    tau,
+    tau_ref_epoch=58849,
+    tolerance=1e-9,
+    max_iter=100,
+    use_c=True,
+    use_gpu=False,
+):
+
+    print("hi Farrah!!")
 
     n_orbs = np.size(sma)  # num sets of input orbital parameters
     n_dates = np.size(epochs)  # number of dates to compute offsets and vz
-
 
     # Necessary for _calc_ecc_anom, for now
     if np.isscalar(epochs):  # just in case epochs is given as a scalar
@@ -58,21 +71,39 @@ def times2trueanom_and_eccanom(sma, epochs, mtot, ecc, tau, tau_ref_epoch=58849,
     # # compute mean anomaly (size: n_orbs x n_dates)
     manom = tau_to_manom(epochs[:, None], sma, mtot, tau, tau_ref_epoch)
     # compute eccentric anomalies (size: n_orbs x n_dates)
-    eanom = _calc_ecc_anom(manom, ecc_arr, tolerance=tolerance, max_iter=max_iter, use_c=use_c, use_gpu=use_gpu)
+    eanom = _calc_ecc_anom(
+        manom,
+        ecc_arr,
+        tolerance=tolerance,
+        max_iter=max_iter,
+        use_c=use_c,
+        use_gpu=use_gpu,
+    )
 
     # compute the true anomalies (size: n_orbs x n_dates)
     # Note: matrix multiplication makes the shapes work out here and below
-    tanom = 2.*np.arctan(np.sqrt((1.0 + ecc)/(1.0 - ecc))*np.tan(0.5*eanom))
+    tanom = 2.0 * np.arctan(np.sqrt((1.0 + ecc) / (1.0 - ecc)) * np.tan(0.5 * eanom))
 
[…]
