### 2024-07-22 a0b603f
add minimal unit test for reflected-light functionality
