### 2024-07-22 04e8638
add if name == main to test_brightness
