### 2024-07-24 017c8ce
modify read_input to read in brightness vals
