### 2024-07-24 03ea984
update todos in test_brightness
