### 2024-07-24 64045dd
test system for brightness calc: gj 504 -> beta pic
