### 2024-07-25 55ab594
Label objects separately and set option to centre plot on star
