### 2024-07-29 a75e461
add ability to fit astrometric jitter
