+++ orbitize/sampler.py
@@ -89,9 +89,9 @@ class Sampler(abc.ABC):
         seppa_indices = self.system.all_seppa
 
         # compute lnlike
+        # NOTE: this won't work if we're fitting more than just astrometry. This is a hack for betelgeuse.
         if "sigma_ast" in self.system.param_idx:
-            errs = np.sqrt(errs**2 + params[self.system.param_idx["sigma_ast"]])
-
+            jitter = np.ones_like(model) * params[self.system.param_idx["sigma_ast"]]
 
         lnlikes = self.lnlike(
             data, errs, corrs, model, jitter, seppa_indices, chi2_type=self.chi2_type
