### 2024-07-29 70c3ef2
fix jitter normalization for hipparcos chi2 calc
