### 2024-07-29 103023d
need to sum jitters to make normalization work
