### 2024-07-29 9b9055e
Added Brightness_calculation
