### 2024-07-30 81f8388
fixed bug returning brightness
