### 2024-08-05 ec07520
Plot of betaPicB data
