### 2024-08-05 4c3f6cf
add brightness to return stms of comp_all_orbits & todo for farrah
