### 2024-08-05 37bbb19
add test_compute_posteriors
