### 2024-08-05 9dcd884
Added brightness as a model prediction
