+++ orbitize/system.py
@@ -667,6 +667,10 @@ class System(object):
                 model[self.rv[body_num], 1] = np.nan
 
             # TODO (farrah): add brightness to model here (use RV block above as template)
+            # Brightness
+            if len(self.brightness[body_num]) > 0:
+                model[self.brightness[body_num], 0] = brightness[self.brightness[body_num], body_num, :]
+                model[self.brightness[body_num], 1] = np.nan
             # (assume self.brightness is array of indices of epochs with brightness measurements)
 
 
