### 2024-08-12 f3ba63f
Fix issue where sep/pa astrometry was being plotted as ra/dec
