### 2024-08-12 717eed8
extend hipparcos unit test to cover situation in #370
