+++ tests/test_hipparcos.py
@@ -258,6 +258,9 @@ def test_save_load_2021():
     myResults = results.Results()
     myResults.load_results(filename)
 
+    filename = "tmp3.hdf5"
+    myResults.results.saveresults(filename)
+
     os.system("rm tmp*.hdf5")
 
 
