### 2024-08-19 222dec7
add pretty visual to test for farrah
