### 2024-08-19 207a032
remove temporary install instructions now that I have access to pypi again *bless*
