+++ docs/index.rst
@@ -55,6 +55,10 @@ User Guide:
 Changelog:
 ++++++++++
 
+**3.0.1 (2024-8-19)**
+
+- bugfix for issue #370 involving saving and loading hipparcos.py objects (@sblunt)
+
 **3.0.0 (2024-4-15)**
 
 - implementation of Hipparcos-Gaia catalog of accelerations fitting! (@semaphoreP)
