### 2024-08-19 #6756 @sblunt
Hey David,

Sorry for the super delayed reply! I am still unable to reproduce this error, so let's try a few more things, a la throw spaghetti at the wall and see what sticks. Here are my suggestions:

1) update astropy and orbitize (I got access to my pypi account back, so you should be able to just `pip install orbitize --upgrade`)
2) It looks like the part of the code that's failing is just reading in a csv file (line 298 of gaia.py). You should have that csv file in orbitize.DATADIR (but make sure, and also open it up to make sure it's not corrupted or anything), so I'd recommend trying to run just this snippet of code and seeing what happens:

```
import os
from orbitize import DATADIR
from astropy.io.ascii import read

gost_filepath = os.path.join(DATADIR, "gaia_edr3_betpic_epochs.csv")

gost_dat = read(gost_filepath, converters={"*": [int, float, bytes]})
```
