### 2024-08-19 #6756 @sblunt
In particular, make sure your version of `gaia_edr3_betpic_epochs.csv` in `orbitize.DATADIR` looks the same as this one: https://github.com/sblunt/orbitize/blob/main/orbitize/example_data/gaia_edr3_betpic_epochs.csv
