### 2024-08-21 #6756 @DTCupcakes
> @DTCupcakes I take it you are recommending for publication then?

Yep
