### 2024-08-21 #6756 @ivastar
@sblunt awesome! I am reviewing the paper text. I'll post some comments shortly. And then we will do the closeout. Final stretch!
