### 2024-08-21 96ad737
fixing some formatting bugs in joss paper
