+++ paper/paper.md
@@ -119,7 +119,7 @@ release of version 1.0 [@Blunt:2020].
 The orbital parameters of directly-imaged planets and binary stars can tell us about
 their present-day dynamics and formation histories [@Bowler:2016], as well as about 
 their inherent physical characteristics (particularly mass, generally called "dynamical 
-mass" when derived from orbital constraints, e.g. [@Brandt:2021], [@Lacour:2021]). 
+mass" when derived from orbital constraints, e.g. @Brandt:2021, @Lacour:2021). 
 
 `orbitize!` is used widely in the exoplanet imaging and binary star communities for 
 translating astrometric data to information about eccentricities [@Bowler:2020], obliquities [@Bryan:2020], 
@@ -130,7 +130,7 @@ Each new released version of the `orbitize!` source code is automatically archiv
 # Major features added since v1
 
 For a detailed overview of the `orbitize!` API, core functionality (including information 
-about our Kepler solver), and initial verification, we refer readers to [@Blunt:2020]. 
+about our Kepler solver), and initial verification, we refer readers to @Blunt:2020. 
 This section lists major new features that have been added to the 
 code since the release of version 1.0 and directs readers to more information about each.
 A complete descriptive list of modifications to the code is maintained in our 
@@ -139,32 +139,32 @@ A complete descriptive list of modifications to the code is maintained in our
 Major new features since v1 include:
 
 1. The ability to jointly fit radial velocity (RV) timeseries, both RVs of the secondary 
-    companion (see Section 3 of [Blunt:2023a]) and RVs of the primary
+    companion (see Section 3 of @Blunt:2023a) and RVs of the primary
     star. RVs of the primary star can either be passed into `orbitize!` directly (see the [radial velocity tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/RV_MCMC_Tutorial.html)), or fit separately and passed in as prior
     information (see the [non-orbitize! posteriors as priors tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/Using_nonOrbitize_Posteriors_as_Priors.html).)
 
 2. The ability to jointly fit absolute astrometry of the primary star. `orbitize!` can fit
-    the Hipparcos-Gaia catalog of accelerations [@Brandt:2021] (see the [HGCA tutorial](https://github.com/sblunt/orbitize/blob/v3/docs/tutorials/HGCA_tutorial.ipynb)), as well as Hipparcos intermediate astrometric data and Gaia 
-    astrometry, following [@Nielsen:2020] (see the [Hipparcos IAD tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/Hipparcos_IAD.html)). It can also handle arbitrary absolute astrometry (Sarah to add tutorial link).
+    the Hipparcos-Gaia catalog of accelerations (@Brandt:2021; see the [HGCA tutorial](https://github.com/sblunt/orbitize/blob/v3/docs/tutorials/HGCA_tutorial.ipynb)), as well as Hipparcos intermediate astrometric data and Gaia 
+    astrometry, following @Nielsen:2020 (see the [Hipparcos IAD tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/Hipparcos_IAD.html)). It can also handle arbitrary absolute astrometry (see the [Fitting Arbitrary Astrometry tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/abs_astrometry.html)).
 
-3. In addition to the MCMC and OFTI posterior computation algorithms documented in [@Blunt:2020], 
-    `orbitize!` version 3 also implements a nested sampling backend, via `dynesty` [@Speagle:2020] 
[…]
