### 2024-08-21 #6756 @sblunt
I just made a few small changes to the paper (fixing formatting things). Can I
