### 2024-08-21 d1636c1
two more typo fixes
