### 2024-08-22 #6756 @ivastar
@sblunt a handful of editorial comments, please review. Let me know if you have questions and let me know when the changes are merged.

Lines 23, 24, 45 and throughout: timeseries --> time series, I think the accepted way to write it is two words
Line 26: expansions --> enhancements
Line 33: used widely --> widely used
Line 33: for translating --> to translate
Line 40, 42: readers --> the reader
Line 42: about each --> about each of them
Line 42: complete --> full
Line 44 and below: the titles of some tutorials are capitalized while others are not. Maybe capitalize all?
Line 63: Keplerians -- > Keplerian orbits? or functions?
Line 71 and 78: Jupyter should be capitalized
Line 76: automatically run -- > run automatically
Line 82: open source doesn't have a hypthen
Lune 84: since --> as or because
Line 91: We recommend users --> We recommend that users
Line 93: orbit fitting, without a hyphen
Line 93: for which treatment -- > for which the treatment
Line 95: which correlated error treatment -- > for which the treatment of correlated errors
Line 99: direct --> refer

The section of lines 45-49 is a bit hard to read, maybe something like this?

The functionality to jointly fit the radial velocity (RV) time series for the primary star together and the secondary companion (refer to section 3 of Blunt et al. (2023)). For the primary star, the RV data can either be directly input into orbitize! (as explained in the radial velocities tutorial) or be fitted separately and then used as priors (as detailed in the non-orbitize! posteriors as priors tutorial).

A comment on the bibliography: our compiler does not parse the \aj, \apj, etc. commands, so please write out the names of the journals instead like \aj --> Astronomical Journal, instead.

Finally, I take your point that the authors of the previous paper did not contribute to the current version but the code they wrote is still in the codebase so it would be nice to at least acknowledge their contributions. I recommend adding them to the acknowledgements section.
