### 2024-08-26 89a3a58
plotting tweaks for corner plot
