+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Blunt:2020,
        author = {{Blunt}, Sarah and {Wang}, Jason J. and {Angelo}, Isabel and {Ngo}, Henry and {Cody}, Devin and {De Rosa}, Robert J. and {Graham}, James R. and {Hirsch}, Lea and {Nagpal}, Vighnesh and {Nielsen}, Eric L. and {Pearce}, Logan and {Rice}, Malena and {Tejada}, Roberto},
         title = "{orbitize!: A Comprehensive Orbit-fitting Software Package for the High-contrast Imaging Community}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {Open source software, Orbit determination, 1866, 1175, Astrophysics - Earth and Planetary Astrophysics},
          year = 2020,
         month = mar,
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Thompson:2023,
        author = {{Thompson}, William and {Lawrence}, Jensen and {Blakely}, Dori and {Marois}, Christian and {Wang}, Jason and {Giordano}, Mos{\'e} and {Brandt}, Timothy and {Johnstone}, Doug and {Ruffio}, Jean-Baptiste and {Ammons}, S. Mark and {Crotts}, Katie A. and {Do {\'O}}, Clarissa R. and {Gonzales}, Eileen C. and {Rice}, Malena},
         title = "{Octofitter: Fast, Flexible, and Accurate Orbit Modeling to Detect Exoplanets}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {Exoplanet detection methods, Bayesian statistics, Direct imaging, 489, 1900, 387, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2023,
         month = oct,
@@ -41,7 +41,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Brandt:2021,
        author = {{Brandt}, Timothy D. and {Dupuy}, Trent J. and {Li}, Yiting and {Brandt}, G. Mirek and {Zeng}, Yunlin and {Michalik}, Daniel and {Bardalez Gagliuffi}, Daniella C. and {Raposo-Pulido}, Virginia},
         title = "{orvara: An Efficient Code to Fit Orbits Using Radial Velocity, Absolute, and/or Relative Astrometry}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {Orbit determination, Keplerian orbit, Astronomy software, Exoplanet astronomy, Astrometric exoplanet detection, Radial velocity, 1175, 884, 1855, 486, 2130, 1332, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2021,
         month = nov,
@@ -62,7 +62,7 @@ archivePrefix = {arXiv},
 @ARTICLE{ONeil:2019,
        author = {{O'Neil}, K. Kosmo and {Martinez}, G.~D. and {Hees}, A. and {Ghez}, A.~M. and {Do}, T. and {Witzel}, G. and {Konopacky}, Q. and {Becklin}, E.~E. and {Chu}, D.~S. and {Lu}, J.~R. and {Matthews}, K. and {Sakai}, S.},
         title = "{Improving Orbit Estimates for Incomplete Orbits with a New Approach to Priors: with Applications from Black Holes to Planets}",
-      journal = {\aj},
[…]
