+++ paper/paper.md
@@ -1,9 +1,9 @@
 ---
-title: 'orbitize! v3: Orbit-fitting for the High-contrast Imaging Community'
+title: 'orbitize! v3: Orbit fitting for the High-contrast Imaging Community'
 tags:
   - Python
   - astronomy
-  - orbit-fitting
+  - Orbit fitting
   - exoplanets
   - high-contrast imaging
 authors:
@@ -106,13 +106,13 @@ bibliography: paper.bib
 # Summary
 
 `orbitize!` is a package for Bayesian modeling of the orbital parameters of resolved binary 
-objects from timeseries measurements. It was developed with the needs of the high-contrast
+objects from time series measurements. It was developed with the needs of the high-contrast
 imaging community in mind, and has since also become widely used in the binary star community.
-A generic `orbitize!` use case involves translating relative astrometric timeseries, optionally 
-combined with radial velocity or astrometric timeseries, into a set of derived orbital posteriors.
+A generic `orbitize!` use case involves translating relative astrometric time series, optionally 
+combined with radial velocity or astrometric time series, into a set of derived orbital posteriors.
 
 This paper is published alongside the release of `orbitize!` version 3.0, which 
-has seen significant expansions in functionality and accessibility since the 
+has seen significant enhancements in functionality and accessibility since the 
 release of version 1.0 [@Blunt:2020].
 
 # Statement of need
@@ -122,8 +122,7 @@ their present-day dynamics and formation histories [@Bowler:2016], as well as ab
 their inherent physical characteristics (particularly mass, generally called "dynamical 
 mass" when derived from orbital constraints, e.g. @Brandt:2021, @Lacour:2021). 
 
-`orbitize!` is used widely in the exoplanet imaging and binary star communities for 
-translating astrometric data to information about eccentricities [@Bowler:2020], obliquities [@Bryan:2020], 
+`orbitize!` is widely used in the exoplanet imaging and binary star communities to translate astrometric data to information about eccentricities [@Bowler:2020], obliquities [@Bryan:2020], 
 dynamical masses [@Lacour:2021], and more. 
 
 Each new released version of the `orbitize!` source code is automatically archived on Zenodo [@orbitize].
@@ -131,71 +130,69 @@ Each new released version of the `orbitize!` source code is automatically archiv
 # Major features added since v1
 
 For a detailed overview of the `orbitize!` API, core functionality (including information 
-about our Kepler solver), and initial verification, we refer readers to @Blunt:2020. 
+about our Kepler solver), and initial verification, we refer the reader to @Blunt:2020. 
 This section lists major new features that have been added to the 
-code since the release of version 1.0 and directs readers to more information about each.
-A complete descriptive list of modifications to the code is maintained in our 
+code since the release of version 1.0 and directs the reader to more information about each of them.
+A full descriptive list of modifications to the code is maintained in our 
 [changelog](https://orbitize.readthedocs.io/en/latest/#changelog).
 
 Major new features since v1 include:
 
-1. The ability to jointly fit radial velocity (RV) timeseries, both RVs of the secondary 
-    companion (see Section 3 of @Blunt:2023a) and RVs of the primary
-    star. RVs of the primary star can either be passed into `orbitize!` directly (see the [radial velocity tutorial](https://orbitize.readthedocs.io/en/latest/tutorials/RV_MCMC_Tutorial.html)), or fit separately and passed in as prior
[…]
