### 2024-08-27 #6756 @ivastar
One minor typo: on line 46 "together and" --> "together with", sorry, I realize that was an error in my suggestion

After that, please tag a version and make a Zenodo release. Please make sure that the Zenodo license is the same as the repo license (common issue). Also, Zenodo archive should have the same title and authors list as the paper.
