### 2024-08-27 051da82
add eric grant info & middle initial
