### 2024-08-28 b3ab054
more plotting tweaks
