+++ orbitize/plot.py
@@ -74,24 +74,24 @@ def plot_corner(results, param_list=None, **corner_kwargs):
     default_labels = {
 
         "sma": "$a_{{B}}$ [au]",
-        "ecc": "$ecc_{{B}}$",
-        "inc": "$inc_{{B}}$ [$^\\circ$]",
+        "ecc": "$ecc$",
+        "inc": "$i$ [$^\\circ$]",
         "aop": "$\\omega_{{B}}$ [$^\\circ$]",
-        "pan": "$\\Omega_{{B}}$ [$^\\circ$]",
-        "tau": "$\\tau_{{B}}$",
+        "pan": "$\\Omega$ [$^\\circ$]",
+        "tau": "$\\tau$",
         "tp": "$T_{{\\mathrm{{P}}}}$",
         "plx": "$\\pi$ [mas]",
         "gam": "$\\gamma$ [km/s]",
         "sig": "$\\sigma$ [km/s]",
         "mtot": "$M_T$ [M$_{{\\odot}}$]",
-        "m0": "$M_{{A}}$ [M$_{{\\odot}}$]",
-        "m": "$M_{{B}}$ [M$_{{\\odot}}$]",
+        "m0": "$M_{{a}}$ [M$_{{\\odot}}$]",
+        "m": "$M_{{b}}$ [M$_{{\\odot}}$]",
         "pm_ra": "$\\mu_{{\\alpha^{{*}}}}$ [mas/yr]",
         "pm_dec": "$\\mu_{{\\delta}}$ [mas/yr]",
         "alpha0": "$\\alpha^{{*}}_{{0}}$ [mas]",
         "delta0": "$\\delta_0$ [mas]",
         # "m": "$M_{0}$ [M$_{{\\rm Jup}}$]",
-        "per": "$P_{{B}}$ [yr]",
+        "per": "$P$ [yr]",
         "K": "$K_{0}$ [km/s]",
         "x": "$X_{0}$ [AU]",
         "y": "$Y_{0}$ [AU]",
