### 2024-09-03 4432d91
fix error message that was printing incorrectly but not affecting code
