### 2024-09-06 e38fed0
add in rob's empirical var correction for type 1 sol
