### 2024-09-09 #6756 @sblunt
Sorry for the delay! I just released a new version, 3.1.0, with the JOSS paper branch merged (and the last typo fixed). Here’s the zenodo DOI: https://zenodo.org/records/13737664. I think everything is copacetic, but let me know if I need to make any other changes.
