### 2024-09-10 aa133fc
incorporate option for ad-hoc brandt+ 23 correction to iad
