### 2024-09-11 #6756 @sblunt
Ah yes, sorry about that. Just changed the zenodo license.
