### 2024-09-12 #6756 @ivastar
@sblunt seems like there are two Brandt:2021 references in the paper.bib file. I am not sure if only one or both are actually cited in the paper and which one it is so don't feel like I can make the change. Can you please review the bib and the paper and resolve the conflict?
