### 2024-09-13 74c0ded
resolve bib conflicts
