+++ paper/paper.bib
@@ -38,7 +38,7 @@ archivePrefix = {arXiv},
 
 
 
-@ARTICLE{Brandt:2021,
+@ARTICLE{orvara,
        author = {{Brandt}, Timothy D. and {Dupuy}, Trent J. and {Li}, Yiting and {Brandt}, G. Mirek and {Zeng}, Yunlin and {Michalik}, Daniel and {Bardalez Gagliuffi}, Daniella C. and {Raposo-Pulido}, Virginia},
         title = "{orvara: An Efficient Code to Fit Orbits Using Radial Velocity, Absolute, and/or Relative Astrometry}",
       journal = {The Astronomical Journal},
