### 2024-09-16 #6756 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
