### 2024-09-17 1863981
add todo for farrah in read input test
