+++ tests/test_brightness.py
@@ -40,7 +40,6 @@ def test_brightness_calculation():
     )
 
     ra, dec, vz, brightness = test_system.compute_all_orbits(params)
-    # TODO (farrah): make plot of brightness vs time
 
     plt.figure()
     plt.scatter(times, brightness)
@@ -51,20 +50,20 @@ def test_brightness_calculation():
 
 def test_read_input_with_brightness():
 
-    # TODO (farrah): use code above as inspiration to read in a csv file with a brightness column
     num_secondary_bodies = 1
 
-    # input_file = os.path.join(DATADIR, "GJ504.csv")
-    input_file = os.path.join(DATADIR, "betaPic.csv")
+    input_file = os.path.join(DATADIR, "reflected_light_example.csv")
 
     data_table = read_input.read_file(input_file)
 
     times = data_table["epoch"].value
     brightness_values = data_table["brightness"].value
 
-    # Do we need the rest of this? since the values for time and brightness are given
+    print(data_table)
 
-    print("hello! :D ")
+    # TODO (Farrah): add a test that asserts the brightness column of the data table is 
+    # what you expect (hint: check in the reflected_light_example.csv to see what
+    # the brightness values should be
 
 
 def test_compute_posteriors():
@@ -113,7 +112,5 @@ def test_compute_posteriors():
 
 if __name__ == "__main__":
     # test_brightness_calculation()
-    # test_read_input_with_brightness()
-    test_compute_posteriors()
-
-    # Test commit
+    test_read_input_with_brightness()
+    # test_compute_posteriors()
