### 2024-09-20 2ca8419
cleaned up test brightness
