+++ tests/test_brightness.py
@@ -8,6 +8,8 @@ import os
 import numpy as np
 import matplotlib.pyplot as plt
 
+print('hello')
+
 
 def test_brightness_calculation():
 
